+++ internal/pkg/cri/cri.go
@@ -1,0 +1,9 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+// Package cri provides minimal CRI client
+package cri
+
+// This client is based on k8s version in https://github.com/kubernetes/kubernetes/tree/master/pkg/kubelet/remote,
+// but it doesn't depend on k8s libs.
